move common elements into core-ui including IGenericChartProps and ChartTypes
