+++ libs/error-analysis/src/lib/ErrorAnalysisDashboard/ErrorAnalysisDashboard.tsx
@@ -13,7 +13,8 @@ import {
   IExplanationModelMetadata,
   ModelTypes,
   WeightVectorOption,
-  WeightVectors
+  WeightVectors,
+  IGenericChartProps
 } from "@responsible-ai/core-ui";
 import {
   DatasetExplorerTab,
@@ -21,7 +22,8 @@ import {
 } from "@responsible-ai/interpret";
 import { localization } from "@responsible-ai/localization";
 import { ModelMetadata } from "@responsible-ai/mlchartlib";
-import _, { Dictionary } from "lodash";
+import { IGlobalExplanationProps } from "@responsible-ai/interpret";
+import _ from "lodash";
 import * as memoize from "memoize-one";
 import {
   IPivotItemProps,
@@ -59,6 +61,7 @@ import {
 } from "./ErrorAnalysisEnums";
 import { ErrorCohort, ErrorDetectorCohortSource } from "./ErrorCohort";
 import { IErrorAnalysisDashboardProps } from "./Interfaces/IErrorAnalysisDashboardProps";
+import { IErrorAnalysisDashboardState } from "./Interfaces/IErrorAnalysisDashboardState";
 import {
   createInitialMatrixAreaState,
   createInitialMatrixFilterState,
@@ -70,77 +73,6 @@ import {
   ITreeViewRendererState,
   createInitialTreeViewState
 } from "./TreeViewState";
-
-export interface IErrorAnalysisDashboardState {
-  activeGlobalTab: GlobalTabKeys;
-  cohorts: ErrorCohort[];
-  customPoints: Array<{ [key: string]: any }>;
-  viewType: ViewTypeKeys;
-  jointDataset: JointDataset;
-  modelMetadata: IExplanationModelMetadata;
-  modelChartConfig?: IGenericChartProps;
-  dataChartConfig?: IGenericChartProps;
-  whatIfChartConfig?: IGenericChartProps;
-  dependenceProps?: IGenericChartProps;
-  globalImportanceIntercept: number[];
-  globalImportance: number[][];
-  importances: number[];
-  isGlobalImportanceDerivedFromLocal: boolean;
-  sortVector?: number[];
-  editingCohortIndex?: number;
-  mapShiftErrorAnalysisOption: ErrorAnalysisOptions;
-  openInfoPanel: boolean;
-  openCohortListPanel: boolean;
-  openEditCohort: boolean;
-  openFeatureList: boolean;
-  openMapShift: boolean;
-  openSaveCohort: boolean;
-  openShiftCohort: boolean;
-  openWhatIf: boolean;
-  predictionTab: PredictionTabKeys;
-  selectedCohort: ErrorCohort;
-  selectedWhatIfIndex: number | undefined;
-  editedCohort: ErrorCohort;
-  baseCohort: ErrorCohort;
-  selectedFeatures: string[];
-  treeViewState: ITreeViewRendererState;
-  matrixAreaState: IMatrixAreaState;
-  matrixFilterState: IMatrixFilterState;
-  errorAnalysisOption: ErrorAnalysisOptions;
-  selectedWeightVector: WeightVectorOption;
-  weightVectorOptions: WeightVectorOption[];
-  weightVectorLabels: Dictionary<string>;
-}
-
-interface IGlobalExplanationProps {
-  globalImportanceIntercept: number[];
-  globalImportance: number[][];
-  isGlobalImportanceDerivedFromLocal: boolean;
-}
-
-export enum ChartTypes {
-  Scatter = "scatter",
-  Bar = "histogram",
-  Box = "box"
-}
-
-export interface IGenericChartProps {
-  chartType: ChartTypes;
-  xAxis?: ISelectorConfig;
-  yAxis?: ISelectorConfig;
-  colorAxis?: ISelectorConfig;
-  selectedCohortIndex?: number;
-}
-
-export interface ISelectorConfig {
-  property: string;
-  index?: number;
-  options: {
-    dither?: boolean;
-    // this is only used in the ambiguous case of numeric values on color axis for scatter chart, when binned or unbinned are valid
-    bin?: boolean;
-  };
-}
 
 export class ErrorAnalysisDashboard extends React.PureComponent<
   IErrorAnalysisDashboardProps,
