--- libs/interpret/src/lib/MLIDashboard/ChartTypes.tsx
@@ -1,9 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-export enum ChartTypes {
-  Scatter = "scatter",
-  Histogram = "histogram",
-  Box = "box",
-  Bar = "bar"
-}
