+++ libs/interpret/src/lib/MLIDashboard/Controls/DatasetExplorerTab/SidePanel.tsx
@@ -18,9 +18,9 @@ import {
 } from "office-ui-fabric-react";
 import React from "react";
 
-import { ChartTypes } from "../../ChartTypes";
+import { ChartTypes } from "@responsible-ai/core-ui";
 import { FabricStyles } from "../../FabricStyles";
-import { IGenericChartProps } from "../../IGenericChartProps";
+import { IGenericChartProps } from "@responsible-ai/core-ui";
 import { InteractiveLegend } from "../InteractiveLegend/InteractiveLegend";
 
 import { datasetExplorerTabStyles } from "./DatasetExplorerTab.styles";
