+++ libs/interpret/src/lib/MLIDashboard/Controls/DatasetExplorerTab/buildCustomData.ts
@@ -7,8 +7,8 @@ import {
   Cohort
 } from "@responsible-ai/core-ui";
 
-import { ChartTypes } from "../../ChartTypes";
-import { IGenericChartProps } from "../../IGenericChartProps";
+import { ChartTypes } from "@responsible-ai/core-ui";
+import { IGenericChartProps } from "@responsible-ai/core-ui";
 
 export function buildCustomData(
   jointData: JointDataset,
