+++ libs/interpret/src/lib/MLIDashboard/Controls/GlobalOnlyChart/GlobalOnlyChart.tsx
@@ -9,7 +9,7 @@ import { localization } from "@responsible-ai/localization";
 import { IDropdownOption, Icon, Slider, Text } from "office-ui-fabric-react";
 import React from "react";
 
-import { ChartTypes } from "../../ChartTypes";
+import { ChartTypes } from "@responsible-ai/core-ui";
 import { FeatureKeys } from "../../SharedComponents/IBarChartConfig";
 import { FeatureImportanceBar } from "../FeatureImportanceBar/FeatureImportanceBar";
 import { globalTabStyles } from "../GlobalExplanationTab/GlobalExplanationTab.styles";
