+++ libs/interpret/src/lib/MLIDashboard/Interfaces/IGlobalExplanationProps.ts
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export interface IGlobalExplanationProps {
+  globalImportanceIntercept: number[];
+  globalImportance: number[][];
+  isGlobalImportanceDerivedFromLocal: boolean;
+}
