+++ libs/model-assessment/src/index.ts
@@ -2,4 +2,3 @@
 // Licensed under the MIT License.
 
 export * from "./lib/ModelAssessmentDashboard/ModelAssessmentDashboard";
-export * from "./lib/ModelAssessmentDashboard/Interfaces/IModelAssessmentDashboardProps";
